third_party/blink/web_tests/platform/generic/fast/loader/scroll-position-restored-on-reload-at-load-event-expected.txt

Test ensures that scrollingElement.scrollTop/Left properties are available by the time DOMContentLoaded event fires.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.scrollingElement.scrollTop is 2000
PASS document.scrollingElement.scrollLeft is 1000
PASS successfullyParsed is true

TEST COMPLETE


    
